+++ lib/dart_quill.dart
@@ -4,14 +4,97 @@ export 'src/assets/assets.dart' show QuillAssets, quillSnowCss;
 export 'src/core/initialization.dart' show initializeQuill;
 export 'src/core/quill.dart';
 export 'src/core/logger.dart' show DebugLevel, logger, setLoggerLevel;
+export 'src/core/emitter.dart' show Emitter, EmitterEvents, EmitterSource;
+export 'src/core/module.dart' show Module;
+export 'src/core/selection.dart'
+    show Bounds, NativePosition, NormalizedNativeRange, Range;
 export 'src/core/theme.dart';
+export 'src/platform/dom.dart';
+
+export 'src/blots/abstract/blot.dart'
+    show
+        Blot,
+        BlotPredicate,
+        EmbedBlot,
+        LeafBlot,
+        ParentBlot,
+        Registry,
+        RegistryEntry,
+        Scope,
+        ScrollBlot;
+export 'src/blots/block.dart' show Block, BlockEmbed;
+export 'src/blots/break.dart' show Break;
+export 'src/blots/container.dart' show Container;
+export 'src/blots/cursor.dart' show Cursor;
+export 'src/blots/embed.dart' show Embed, EmbedContextRange;
+export 'src/blots/inline.dart' show Inline, InlineBlot;
+export 'src/blots/scroll.dart' show Scroll;
+export 'src/blots/text.dart' show TextBlot;
+
+export 'src/formats/abstract/attributor.dart'
+    show Attributor, AttributorStore, ClassAttributor, StyleAttributor;
+export 'src/formats/align.dart';
+export 'src/formats/background.dart' hide config;
+export 'src/formats/blockquote.dart';
+export 'src/formats/bold.dart';
+export 'src/formats/code.dart';
+export 'src/formats/color.dart';
+export 'src/formats/direction.dart';
+export 'src/formats/font.dart';
+export 'src/formats/formula.dart';
+export 'src/formats/header.dart';
+export 'src/formats/image.dart';
+export 'src/formats/indent.dart';
+export 'src/formats/italic.dart';
+export 'src/formats/link.dart';
+export 'src/formats/list.dart';
+export 'src/formats/script.dart';
+export 'src/formats/size.dart' hide config;
+export 'src/formats/strike.dart';
+export 'src/formats/table.dart';
+export 'src/formats/underline.dart';
+export 'src/formats/video.dart';
+
+export 'src/modules/clipboard.dart'
+    show
+        ATTRIBUTE_ATTRIBUTORS,
+        Clipboard,
+        ClipboardOptions,
+        Matcher,
+        STYLE_ATTRIBUTORS,
+        Selector;
+export 'src/modules/history.dart'
+    show History, HistoryOptions, Stack, StackItem;
+export 'src/modules/input.dart' show Input, InputOptions;
+export 'src/modules/keyboard.dart'
+    show
+        BindingHandler,
+        BindingObject,
+        Context,
+        DefaultBindingHandler,
+        Keyboard,
+        KeyboardOptions,
+        NormalizedBinding,
+        SHORTKEY,
+        deleteRange,
+        isMacPlatform;
 export 'src/modules/syntax.dart';
 export 'src/modules/ui_node.dart';
 export 'src/modules/toolbar.dart';
 export 'src/modules/table.dart' show Table, TableOptions;
+export 'src/modules/table_embed.dart' show TableEmbed;
 export 'src/modules/image_resize.dart'
     show ImageResize, ImageResizeOptions, ImageWrap;
 export 'src/modules/uploader.dart';
+export 'src/modules/normalize_external_html/index.dart'
+    show NormalizeExternalHTML;
+
+export 'src/themes/base.dart'
+    show ALIGNS, BaseTheme, BaseTooltip, COLORS, FONTS, HEADERS, SIZES;
 export 'src/themes/bubble.dart' show BubbleTheme;
 export 'src/themes/snow.dart' show SnowTheme;
+export 'src/ui/icons.dart';
+export 'src/ui/picker.dart' show ColorPicker, IconPicker, Picker;
+export 'src/ui/tabler_icons.dart';
+export 'src/ui/tooltip.dart' show Tooltip, isScrollable;
 export 'src/dependencies/dart_quill_delta/dart_quill_delta.dart';